Modify: PlayingContainer에서 소켓 서버 이벤트 수신 등록 및 콜백 정의
