+++ src/components/Game/Playing/WordInput.jsx
@@ -20,7 +20,7 @@ import {
   thisTurnLeftTimeState,
   thisRoundLeftTimeState,
   currentRoundState,
-  currentPoints,
+  currentPointsState,
   isWordFailState,
   timeoutIdsState
 } from "@/recoil/gameState";
@@ -40,7 +40,7 @@ const WordInput = ({ roundCount, roundTime }) => {
   const [inputWord, setInputWord] = useState("");
   const [isFail, setIsFail] = useRecoilState(isWordFailState);
 
-  const setCurrPoints = useSetRecoilState(currentPoints);
+  const setCurrPoints = useSetRecoilState(currentPointsState);
   const [playerList, setPlayerList] = useRecoilState(playingPlayerListState);
   const [player, setPlayer] = useRecoilState(playingPlayerState);
   const syncPlayerList = useSetRecoilState(syncPlayingPlayerToListState);
@@ -58,11 +58,11 @@ const WordInput = ({ roundCount, roundTime }) => {
   }, [player]);
 
   // clearTimeout
-  useEffect(() => {
-    return () => {
-      timeoutIds?.forEach((id) => clearTimeout(id));
-    };
-  }, [timeoutIds]);
+  // useEffect(() => {
+  //   return () => {
+  //     timeoutIds?.forEach((id) => clearTimeout(id));
+  //   };
+  // }, [timeoutIds]);
 
   const onEnterKeyDown = async (e) => {
     if (e.key !== "Enter") return;
@@ -73,66 +73,66 @@ const WordInput = ({ roundCount, roundTime }) => {
     // 서버로 단어 전송 및 유효 단어 여부 판별 요청
     sendMessage(inputWord);
 
-    // 끝말잇기 실패 시
-    receiveSayWordFail((word) => {
-      const prevInitialCharacter = initialCharacter;
-      setIsFail(true);
-      setInitialCharacter(word);
+    // // 끝말잇기 실패 시
+    // receiveSayWordFail((word) => {
+    //   const prevInitialCharacter = initialCharacter;
+    //   setIsFail(true);
+    //   setInitialCharacter(word);
 
-      const id = setTimeout(() => {
-        setInitialCharacter(prevInitialCharacter);
-        setIsFail(false);
-      }, 1000);
+    //   const id = setTimeout(() => {
+    //     setInitialCharacter(prevInitialCharacter);
+    //     setIsFail(false);
+    //   }, 1000);
 
-      setTimeoutIds([id]);
-    });
+    //   setTimeoutIds([id]);
+    // });
 
-    // 끝말잇기 성공 시
-    receiveSayWordSucceed((data) => {
-      const { word, userIndex, scoreDelta } = data;
+    // // 끝말잇기 성공 시
+    // receiveSayWordSucceed((data) => {
+    //   const { word, userIndex, scoreDelta } = data;
 
-      // 다음 끝말잇기 글자 설정
-      setInitialCharacter((prevChar) => prevChar + word.split("")[word.length - 1]);
-      const inputWordCharacters = inputWord?.split("");
-      const delay = 500; // 0.5초
-      inputWordCharacters.forEach((char, idx) => {
-        const id1 = setTimeout(
-          () => {
-            if (idx !== 0) setInitialCharacter((prevChar) => prevChar + char);
-          },
-          delay * (idx + 1)
-        );
-        timeoutIds.push(id1);
-      });
+    //   // 다음 끝말잇기 글자 설정
+    //   setInitialCharacter((prevChar) => prevChar + word.split("")[word.length - 1]);
+    //   const inputWordCharacters = inputWord?.split("");
+    //   const delay = 500; // 0.5초
+    //   inputWordCharacters.forEach((char, idx) => {
+    //     const id1 = setTimeout(
+    //       () => {
+    //         if (idx !== 0) setInitialCharacter((prevChar) => prevChar + char);
+    //       },
+    //       delay * (idx + 1)
+    //     );
+    //     timeoutIds.push(id1);
+    //   });
 
-      // 득점 저장
-      setCurrPoints(scoreDelta);
-      setPlayer((prev) => {
-        const newRoundScore = [...(prev.roundScore || [])];
-        newRoundScore[currRound] = scoreDelta;
-        return {
-          ...prev,
-          roundScore: newRoundScore,
-          totalScore: prev.totalScore + scoreDelta
-        };
-      });
-      setPlayerList((prevList) => {
-        const newList = [...prevList];
-        const _player = newList[userIndex];
-        _player.totalScore += scoreDelta;
-        return newList;
-      });
+    //   // 득점 저장
+    //   setCurrPoints(scoreDelta);
+    //   setPlayer((prev) => {
+    //     const newRoundScore = [...(prev.roundScore || [])];
+    //     newRoundScore[currRound] = scoreDelta;
+    //     return {
+    //       ...prev,
+    //       roundScore: newRoundScore,
+    //       totalScore: prev.totalScore + scoreDelta
+    //     };
+    //   });
+    //   setPlayerList((prevList) => {
+    //     const newList = [...prevList];
+    //     const _player = newList[userIndex];
+    //     _player.totalScore += scoreDelta;
+    //     return newList;
+    //   });
 
-      const id2 = setTimeout(
-        () => {
-          setInitialCharacter(inputWord?.split("")[inputWord?.length - 1]);
-        },
-        delay * 1.5 * inputWordCharacters?.length
-      );
+    //   const id2 = setTimeout(
+    //     () => {
+    //       setInitialCharacter(inputWord?.split("")[inputWord?.length - 1]);
+    //     },
+    //     delay * 1.5 * inputWordCharacters?.length
+    //   );
 
-      timeoutIds.push(id2);
-      setTimeoutIds(timeoutIds);
-    });
+    //   timeoutIds.push(id2);
+    //   setTimeoutIds(timeoutIds);
+    // });
   };
 
   return (
