+++ src/components/Game/Waiting/WaitingTab.jsx
@@ -3,7 +3,7 @@ import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
 import PropTypes from "prop-types";
 
 import { gameStart, switchReadyState } from "@/services/socket";
-import { roomIdState, isPlayingRoomState } from "@/recoil/roomState";
+import { roomIdState } from "@/recoil/roomState";
 import { waitingPlayerListState, playingPlayerListState } from "@/recoil/userState";
 import { userNameState } from "@/recoil/userState";
 import { SpacingWrapper } from "../Shared/Layout";
@@ -32,24 +32,6 @@ const WaitingTab = ({ isHost, roomId, setIsPlaying }) => {
 
   const onStartGame = () => {
     gameStart();
-    // (room) => {
-    //   const updatedPlayerList = room.usersSequence.map((user, idx) => ({
-    //     id: user.userId,
-    //     score: user.score,
-    //     myTurn: idx === 0
-    //   }));
-    //   // setMyTurnPlayerIndex(room.currentTurnUserIndex);
-    //   setPlayingPlayerList(updatedPlayerList);
-    //   setCurrRound(room.currentRound);
-    //   setRandomWord(room.roundWord);
-    //   setInitialCharacter(room.roundWord[0]);
-    //   setIsPlaying(true);
-    // },
-    // (error) => {
-    //   setModalType("alert");
-    //   setModalMessage(error?.slice(1, -1));
-    //   setIsModalOpen(true);
-    // }
   };
 
   const onUpdateRoomConfig = () => {
