+++ src/recoil/gameState.js
@@ -49,20 +49,15 @@ export const myTurnPlayerIndexState = selector({
     get(playingPlayerListState).findIndex((player) => player.myTurn === true)
 });
 
-// (임시)
+// 끝말잇기 실패 상태
 export const isWordFailState = atom({
   key: "isWordFailState",
   default: false
 });
 
-export const timeoutIdsState = atom({
-  key: "timeoutIdsState",
-  default: []
-});
-
 // 현재 차례 끝말잇기 점수 결과
-export const currentPoints = atom({
-  key: "currentPoints",
+export const currentPointsState = atom({
+  key: "currentPointsState",
   default: 0
 });
